tests/requirements-test.txt
pytest
requests
psycopg2-binary
python-dotenv
